+++ src/App.jsx
@@ -1,11 +1,11 @@
-import { useSnakbarContext } from "./Context/SnakbarContext";
+import { useSnakbar } from "./Context/SnakbarProvider";
 import { NavBar } from "./Components/NavBar";
 import { Snakbar } from "./Components/Snakbar";
 import { Navigation } from "./Components/Navigation/Navigation";
 import { ErrorModel } from "./Components/ErrorModel";
-import { useStatus } from "./Context/LoaderContext";
+import { useStatus } from "./Context/LoaderProvider";
 export default function App() {
-  const { snakbarStatus } = useSnakbarContext();
+  const { snakbarStatus } = useSnakbar();
   const { status } = useStatus();
   return (
     <main>
