--- src/Context/AuthContext/index.js
@@ -1,1 +1,0 @@
-export { AuthProvider, useAuthContext } from "./AuthProvider";
