--- src/Context/LoaderContext/index.js
@@ -1,1 +1,0 @@
-export { LoaderContextProvider, useStatus } from "./LoaderContext";
