+++ src/Context/LoaderProvider/LoaderContext.jsx
@@ -2,7 +2,7 @@ import { createContext, useState, useContext } from "react";
 
 const LoaderContext = createContext();
 
-export const LoaderContextProvider = ({ children }) => {
+export const LoaderProvider = ({ children }) => {
   const [status, setStatus] = useState("IDLE");
   return (
     <LoaderContext.Provider value={{ status: status, setStatus: setStatus }}>
