--- src/Context/SnakbarContext/index.js
@@ -1,1 +1,0 @@
-export { SnakbarContextProvider, useSnakbarContext } from "./SnakbarContext";
