+++ src/components/ErrorModel/ErrorModel.jsx
@@ -1,5 +1,5 @@
 import { Model } from "../Model";
-import { useStatus } from "Context/LoaderContext";
+import { useStatus } from "Context/LoaderProvider";
 import { useLocation, useNavigate } from "react-router-dom";
 export const ErrorModel = () => {
   const { status, setStatus } = useStatus();
