+++ src/pages/Home/Home.jsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from "react";
 import wapperbackground from "assests/images/wall.png";
 import "./home.css";
 import { ProductCard } from "Components/ProductCard";
-import { useRequest } from "utils/request";
-import { useStatus } from "Context/LoaderContext";
+import { useRequest } from "utils/request.hook";
+import { useStatus } from "Context/LoaderProvider";
 import { Link, useNavigate } from "react-router-dom";
 import { Loader } from "Components/Loader";
 import { ErrorModel } from "Components/ErrorModel";
